Add files via upload
Q5. Give a greedy algorithm for making change with
US currency.US coin denominations of (1 cent, 5 cent,10
cent,25 cent,50 cent, and dollar.
1. Problem statement:Trying to give back the change with
largest coins as possible.
2. Runtime analysis:The algorithm need O(n) time complexity.
as we can see, the algorithm has 15 simple lines of
code and 4 loops. and because all of those are Constance
the time complexity is O(n)
3.Proof of correctness
1. Partial Correctness:
a.initialization:Start with 0 or 1cent coins or dollars.
b.Maintenance, having largest coins sorted and the rests
1cent coins. c.Termination:No more coins to change and
ending with an order of largest to smallest coins
2. Termination: At the last iteration, No more coins to
change and largest coins is sorted, and all coins in coin_i
are larger than elements in coin_i-1. Hence coins is sorted
4.solution: code